modified files to fit the demo needs

+++ include/Process.hpp
@@ -41,7 +41,7 @@ private:
   std::string name_;
   std::string created_at_ = Utils::current_timestamp();
   std::vector<std::unique_ptr<IInstruction>> instructions_;
-  std::unordered_map<std::string, int> symbol_table_;
+  std::unordered_map<std::string, uint16_t> symbol_table_;
   int total_instructions_;
   int instruction_pointer_ = 0;
   ProcessState state_;
